#7120 the painter expects a float as line_thickness_, simplify initial computation of line_thickness_ and line_offset_
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@36890 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/rowpainter.h
@@ -104,7 +104,7 @@ private:
 	int const yo_;    // current baseline
 	double x_;
 	int width_;
-	int line_thickness_;
+	float line_thickness_;
 	int line_offset_;
 };
